requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jiter==0.8.2
joblib==1.4.2
nltk==3.9.1
openai==1.65.1
pydantic==2.10.6
pydantic_core==2.27.2
pyphen==0.17.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
textstat==0.7.5
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
youtube-transcript-api==0.6.3
zipp==3.21.0
